add load-forms & add late init of spec
The late initialization is important to allow us to be able to use
functions defined in the same file as a constraint or param.

We initally had functions in the specs, but they are not writable (via
load-form) into a fasl. So we started storing the names as well so we
could write the name into the load-form. This was good but when a
spec is made during macroexpansion the function named probably hasnt
been defined yet, so now we store the name and only find the function
on the first creation of a type from the spec.