extract special-form inference, remove generic funcs, add lang pkg
We dont need the functions to be generic. There can only be 1
implementation per type system and this sets us specify it
directly. Later on we will be moving to using structs and this will
let us type these functions for more efficient dispatch.

infer-if is no longer part of checkmate. That must be provided by the
consuming project

We add the checkmate.lang package which exports the symbols that name
the forms that checkmate can infer. Most are just re-exported CL
symbols.

One major downside right now is that we are storing the infer
functions as symbols as otherwise live recompile doesnt work