enable continuity of unknown type across multiple calls
The summary is too short to explain properly. In short a bunch of
functions now allow the passing of the hash-table that is used to
ensure that the same unknown-type designator maps to the same type.

This means that in the following example the multiple calls to
`find-ttype` will create the same type object for `?a`.

    (let ((unknowns (make-hash-table)))
      (find-ttype 'tables '?a unknowns)
      (find-ttype 'tables '?a unknowns))

or perhaps more usefully

    (let ((unknowns (make-hash-table)))
      (find-ttype 'tables '(function (?a) ?b) unknowns)
      (find-ttype 'tables '(function (?a ?a) boolean) unknowns))